+++ apps/frontend/src/hooks/useGetSession.ts
@@ -1,4 +1,4 @@
-import { useSelector } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
 import { useLocation, useNavigate } from 'react-router-dom';
 import { selectFirstSessionData } from '../features/app/selectors';
 import { useCallback, useEffect, useMemo } from 'react';
@@ -6,21 +6,26 @@ import { useLazyGetSessionQuery } from '../features/app/services';
 import { APP_ROUTES } from '../utils/appRoutes';
 import { isFetchBaseQueryError } from '../utils/isFetchBaseQueryError';
 import { getPathFromEvironment } from '../utils/getDefaultPathFromEnv';
+import { authActions } from '../features/auth/reducer';
+import { selectIsTokenValid } from '../features/auth/selectors';
 
 const redirectTokenError = { data: 'Session ID not provided', status: 401 };
 
 export const useGetSession = () => {
+  const dispatch = useDispatch();
   const { search } = useLocation();
   const navigate = useNavigate();
 
   const session = useSelector(selectFirstSessionData);
+
+  const isChachedSessionValid = useSelector(selectIsTokenValid);
 
   const redirectToken = useMemo(() => new URLSearchParams(search).get('id'), [search]);
 
   const [getSession] = useLazyGetSessionQuery();
 
   const retrieveSession = useCallback(async () => {
-    if (!redirectToken) {
+    if (!redirectToken && !isChachedSessionValid) {
       navigate(APP_ROUTES.UNAUTHORIZED, {
         state: {
           status: redirectTokenError.status,
@@ -29,16 +34,21 @@ export const useGetSession = () => {
       return;
     }
 
-    if (session && session.token) {
-      navigate(getPathFromEvironment());
+    if ((session && session.token) || isChachedSessionValid) {
+      navigate(APP_ROUTES.SELECT_YEAR);
       return;
     }
 
-    const { isError: sessionError, error: sessionErrorMsg } = await getSession({
+    const {
+      isError: sessionError,
+      error: sessionErrorMsg,
+      data,
+    } = await getSession({
       id: redirectToken!,
     });
 
     if (sessionError && isFetchBaseQueryError(sessionErrorMsg)) {
+      dispatch(authActions.clearToken());
       navigate(APP_ROUTES.UNAUTHORIZED, {
         state: {
           status: sessionErrorMsg.status,
@@ -46,9 +56,12 @@ export const useGetSession = () => {
       });
       return;
     }
+    if (data?.token) {
+      dispatch(authActions.setToken(data?.token));
+    }
     navigate(getPathFromEvironment());
     return;
-  }, [getSession, navigate, redirectToken, session]);
+  }, [dispatch, getSession, isChachedSessionValid, navigate, redirectToken, session]);
 
   useEffect(() => {
     retrieveSession();