+++ apps/frontend/src/main.tsx
@@ -1,24 +1,30 @@
 import { createRoot } from 'react-dom/client';
 import { CssBaseline } from '@mui/material';
 import { theme } from '@io-cdc/ui';
-
 import { BrowserRouter } from 'react-router-dom';
 import './index.css';
 import App from './App.tsx';
 import { ThemeProvider } from '@mui/material/styles';
 import { Provider } from 'react-redux';
-import { store } from './features/store.ts';
+import { persistor, store } from './features/store.ts';
 import { ToastProvider } from './contexts/index.ts';
+import { PersistGate } from 'redux-persist/integration/react';
+import { Suspense } from 'react';
+import { RequestLoader } from './components/RequestLoader/index.tsx';
 
 createRoot(document.getElementById('root')!).render(
   <Provider store={store}>
-    <ThemeProvider theme={theme}>
-      <ToastProvider>
-        <CssBaseline />
-        <BrowserRouter>
-          <App />
-        </BrowserRouter>
-      </ToastProvider>
-    </ThemeProvider>
+    <Suspense fallback={<RequestLoader />}>
+      <PersistGate persistor={persistor}>
+        <ThemeProvider theme={theme}>
+          <ToastProvider>
+            <CssBaseline />
+            <BrowserRouter>
+              <App />
+            </BrowserRouter>
+          </ToastProvider>
+        </ThemeProvider>
+      </PersistGate>
+    </Suspense>
   </Provider>,
 );