+++ apps/frontend/CHANGELOG.md
@@ -1,5 +1,23 @@
 # frontend
 
+## 1.0.3
+
+### Patch Changes
+
+- b63a9dd: Handle rehydrate persist store with suspense
+
+## 1.0.2
+
+### Patch Changes
+
+- 856a67d: added session cache
+
+## 1.0.1
+
+### Patch Changes
+
+- 122417c: Handle expired config when years list is empty
+
 ## 1.0.0
 
 ### Major Changes